requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
playwright==1.44.0
pydantic==2.7.1
